+++ spec/helpers/alchemy/admin/tags_helper_spec.rb
@@ -6,74 +6,67 @@ module Alchemy
     let(:tag2) { mock_model(ActsAsTaggableOn::Tag, name: 'abc', count: 1) }
 
     let(:params) do
-      ActionController::Parameters.new({
-        controller: 'alchemy/admin/attachments',
-        action: 'index',
-        use_route: 'alchemy',
-        tagged_with: 'foo'
-      })
+      ActionController::Parameters.new(tagged_with: 'foo')
     end
 
     before do
-      allow(helper).to receive(:params) { params }
-      allow(helper).to receive(:options_from_params) do
-        ActionController::Parameters.new.permit!
+      allow(helper).to receive(:search_filter_params) do
+        params.permit!.merge(controller: 'admin/attachments', action: 'index', use_route: 'alchemy')
       end
     end
 
     describe '#render_tag_list' do
+      subject { helper.render_tag_list('Alchemy::Attachment') }
+
       context "with tagged objects" do
         before { allow(Attachment).to receive(:tag_counts).and_return([tag, tag2]) }
 
         it "returns a tag list as <li> tags" do
-          expect(helper.render_tag_list('Alchemy::Attachment')).to match(/li/)
+          is_expected.to match(/li/)
         end
 
         it "has the tags name in the li's name attribute" do
-          expect(helper.render_tag_list('Alchemy::Attachment')).to match(/li.+name="#{tag.name}"/)
+          is_expected.to match(/li.+name="#{tag.name}"/)
         end
 
         it "has active class if tag is present in params" do
-          expect(helper.render_tag_list('Alchemy::Attachment')).to match(/li.+class="active"/)
+          is_expected.to match(/li.+class="active"/)
         end
 
         it "tags are sorted alphabetically" do
-          expect(helper.render_tag_list('Alchemy::Attachment')).to match(/li.+name="#{tag2.name}.+li.+name="#{tag.name}/)
+          is_expected.to match(/li.+name="#{tag2.name}.+li.+name="#{tag.name}/)
         end
 
         context "with lowercase and uppercase tag names mixed" do
           let(:tag) { mock_model(ActsAsTaggableOn::Tag, name: 'Foo', count: 1) }
 
           it "tags are sorted alphabetically correctly" do
-            expect(helper.render_tag_list('Alchemy::Attachment')).to match(/li.+name="#{tag2.name}.+li.+name="#{tag.name}/)
+            is_expected.to match(/li.+name="#{tag2.name}.+li.+name="#{tag.name}/)
           end
         end
 
         it "output is html_safe" do
-          expect(helper.render_tag_list('Alchemy::Attachment').html_safe?).to be_truthy
+          expect(subject.html_safe?).to be(true)
         end
 
         context "when filter and search params are present" do
           let(:params) do
-            ActionController::Parameters.new({
-              controller: 'alchemy/admin/attachments',
-              action: 'index',
-              use_route: 'alchemy',
+            ActionController::Parameters.new(
               filter: 'foo',
               q: {name_eq: 'foo'}
-            })
+            )
           end
 
           it 'keeps them' do
-            expect(helper.render_tag_list('Alchemy::Attachment')).to match(/filter/)
-            expect(helper.render_tag_list('Alchemy::Attachment')).to match(/name_eq/)
+            is_expected.to match(/filter/)
+            is_expected.to match(/name_eq/)
           end
         end
       end
 
       context "without any tagged objects" do
         it "returns empty string" do
-          expect(helper.render_tag_list('Alchemy::Attachment')).to be_empty
+          is_expected.to be_empty
         end
       end
 
@@ -84,108 +77,64 @@ module Alchemy
       end
     end
 
-    describe '#tag_list_tag_active?' do
-      context "the tag is in params" do
-        it "returns true" do
-          expect(helper.tag_list_tag_active?(tag)).to be_truthy
+    describe "#filtered_by_tag?" do
+      subject { helper.filtered_by_tag?(tag) }
+
+      context 'if the filter list params contains the given tag' do
+        let(:params) do
+          ActionController::Parameters.new(tagged_with: 'foo,bar,baz')
         end
+
+        it { is_expected.to eq(true) }
+      end
+
+      context 'if the filter list params does not contain the given tag' do
+        let(:params) do
+          ActionController::Parameters.new(tagged_with: 'bar,baz')
+        end
+
+        it { is_expected.to eq(false) }
       end
 
       context "params[:tagged_with] is not present" do
         let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy'
-          })
+          ActionController::Parameters.new
         end
 
-        it "returns false" do
-          expect(helper.tag_list_tag_active?(tag)).to be_falsey
-        end
+        it { is_expected.to eq(false) }
       end
     end
 
-    describe "#filtered_by_tag?" do
-      it "should return true if the filterlist contains the given tag" do
-        controller.params[:tagged_with] = "foo,bar,baz"
-        expect(helper.filtered_by_tag?(tag)).to eq(true)
-      end
+    describe "#tags_for_filter" do
+      subject { helper.tags_for_filter(current: tag) }
 
-      context 'if the filterlist does not contain the given tag' do
-        let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy',
-            tagged_with: 'bar,baz'
-          })
-        end
-
-        it "should return false" do
-          expect(helper.filtered_by_tag?(tag)).to eq(false)
-        end
-      end
-    end
-
-    describe "#add_to_tag_filter" do
       context "if params[:tagged_with] is not present" do
         let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy'
-          })
+          ActionController::Parameters.new
         end
 
-        it "should return an Array with the given tag name" do
-          expect(helper.add_to_tag_filter(tag)).to eq(["foo"])
+        it "should return a String with the given tag name" do
+          is_expected.to eq("foo")
         end
       end
 
       context "if params[:tagged_with] contains some tag names" do
         let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy',
-            tagged_with: 'bar,baz'
-          })
+          ActionController::Parameters.new(tagged_with: 'bar,baz')
         end
 
-        it "should return an Array of tag names including the given one" do
-          expect(helper.add_to_tag_filter(tag)).to eq(["bar", "baz", "foo"])
-        end
-      end
-    end
-
-    describe "#remove_from_tag_filter" do
-      context "if params[:tagged_with] is not present" do
-        let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy'
-          })
-        end
-
-        it "should return an empty Array" do
-          expect(helper.remove_from_tag_filter(tag)).to be_empty
+        it "should return a String of tag names including the given one" do
+          is_expected.to eq("bar,baz,foo")
         end
       end
 
-      context "if params[:tagged_with] contains some tag names" do
+      context "if params[:tagged_with] contains current tag name" do
         let(:params) do
-          ActionController::Parameters.new({
-            controller: 'alchemy/admin/attachments',
-            action: 'index',
-            use_route: 'alchemy',
-            tagged_with: 'bar,baz,foo'
-          })
+          ActionController::Parameters.new(tagged_with: 'bar,baz,foo')
         end
 
-        it "should return an Array of tag names without the given one" do
-          expect(helper.remove_from_tag_filter(tag)).to eq(["bar", "baz"])
+        it "should return a String of tag names without the current one" do
+          is_expected.to eq("bar,baz")
         end
       end
     end